requirements.txt
streamlit==1.43.2
langchain-google-genai==2.1.1
langchain-community==0.3.20
langchain-text-splitters==0.3.7
langchain-core==0.3.47
nltk==3.9.1
pypdf==5.4.0
chromadb==0.6.3
python-dotenv==1.0.1